refactor(RestaurantList): 음식점 모델 삭제 후 일반 객체를 이용
MVC 패턴을 위한 코드 작성을 한 것 같아 삭제(객체를 객체답게 사용하지 않았다)
